refactor(pubsub): split test files by source file (1:1 naming convention)
Reorganize the test files in pubsub/ to follow the Go Code Review
Comments recommendation that test files be tightly coupled to their
corresponding source file by name (foo.go ↔ foo_test.go). This makes
the test surface discoverable by walking the source tree in any IDE
or by reading git log/blame at the source level.

Splits the 24 test functions previously scattered across
pubsub_test.go and coverage_test.go into:

  - broker_test.go     (11 tests + testLogger helper + packet type)
                       covers WithLogger, WithId, WithCapacity,
                       Broker.Id/Capacity/String/Topics, and the
                       end-to-end integration scenarios (basic
                       round-trip, fan-out, fan-in, fan-in-out,
                       struct-payload round-trip, concurrent stress)

  - publisher_test.go  (3 tests)
                       covers Publisher.Id/String and Publish
                       (auto-create, capacity-exceeded)

  - subscriber_test.go (10 tests)
                       covers Subscriber.Id/String, lifecycle
                       (Subscribe-after-Close, double-Close, multi-
                       topic Close), WithTimeout (idle + sliding
                       reset), WithChannelSize (overflow), Subscribes
                       (partial-failure, multi-topic happy path)

The two remaining package-level test files are kept by design:
bench_test.go (benchmarks are conventionally grouped by theme, not
source file) and example_test.go (the godoc Example* naming
convention requires this exact filename). leaktest_test.go holds
the TestMain that runs goleak.VerifyTestMain on every test.

Pure reorganization, zero behavior change: go test -race -count=1
./pubsub/... still completes in 2.8 s and coverage remains at
95.7% of statements.
